refactor: separate into PostList and HistoryList
