+++ package.json
@@ -22,6 +22,7 @@
         "@nuxt/test-utils": "^3.12.1",
         "@nuxtjs/color-mode": "^3.4.0",
         "@pinia/nuxt": "^0.5.1",
+        "@pinia/testing": "^0.1.3",
         "@testing-library/vue": "^8.0.3",
         "@unocss/eslint-config": "^0.59.4",
         "@unocss/nuxt": "^0.59.4",
